requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
pytesseract
pdf2image
opencv-python-headless
numpy
requests
pillow
python-multipart
slowapi
limits
python-dotenv